# src/swagger.json
{
	"components": {
...
		"schemas": {
			"FieldErrors": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
...
			"VolunteerError": {
				"properties": {
					"name": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ErrorName"
							},
							{
								"type": "string",
								"enum": [
									"UNSPECIFIED_ERROR"
								]
							}
						]
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"details": {}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PaginationResult_BookClassAllInfo-Array_": {
				"properties": {
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"pageInfo": {
						"properties": {
							"hasPreviousPage": {
								"type": "boolean"
							},
							"hasNextPage": {
								"type": "boolean"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"hasPreviousPage",
							"hasNextPage",
							"page"
						],
						"type": "object"
					},
					"nodes": {
						"items": {
							"$ref": "#/components/schemas/BookClassAllInfo"
						},
						"type": "array"
					}
				},
				"required": [
					"totalCount",
					"pageInfo",
					"nodes"
				],
				"type": "object"
			},
			"BookClubClassError": {
				"properties": {
					"name": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ErrorName"
							},
							{
								"type": "string",
								"enum": [
									"UNSPECIFIED_ERROR"
								]
							}
						]
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"details": {}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AvailableClassRowEntity": {
				"properties": {
					"idclass": {
						"type": "number",
						"format": "double"
					},
					"place": {
						"type": "string"
					},
					"dateReserved": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"dateConcluded": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"folderLink": {
						"type": "string"
					},
					"numEvaluations": {
						"type": "number",
						"format": "double"
					},
					"totalEssays": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"idclass",
					"numEvaluations",
					"totalEssays"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ReserveClassDataEntity": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_VolunteerEntity.idvol_"
					},
					{
						"$ref": "#/components/schemas/Pick_BookClubClassEntity.idclass_"
					}
				]
			},
...
			"PepClassError": {
				"properties": {
					"name": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ErrorName"
							},
							{
								"type": "string",
								"enum": [
									"UNSPECIFIED_ERROR"
								]
							}
						]
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"details": {}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdatePepClassEntity": {
				"$ref": "#/components/schemas/Omit_PepClassEntity.id-or-placeId_"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "palavrasdepaz",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/book-club-class": {
			"get": {
				"operationId": "GetVolunteersAttendanceMetrics",
				"responses": {
					"200": {
						"description": "Successfully generated the metrics",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginationResult_BookClassAllInfo-Array_"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all book-club-class.\n\n(The user must have bookPermissionn, which is checked using JWT)\n Pagination\n Page: ?page=number& (page number)\n Limit: ?limit=number& (data quantity - max=30)",
				"tags": [
					"Book Club Class"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/book-club-class/download/{idclass}": {
			"get": {
				"operationId": "DownloadClassReport",
				"responses": {
					"200": {
						"description": "Successfully downloaded the files",
						"content": {
							"application/json": {
								"schema": {
									"type": "string",
									"format": "byte"
								}
							}
						}
					},
					"404": {
						"description": "Essay not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookClubClassError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "ESSAY_NOT_FOUND",
											"message": "Essay with id {some class id} not found"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Download files of the book club class as a zip\nThe return of this route is a stream (content type: application/octet-stream)\nand the response header include the file as an attachement\n\n(The volunteer must have bookPermission, which is checked using JWT)",
				"tags": [
					"Book Club Class"
				],
				"security": [
					{
						"jwt": [
							"bookPermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "idclass",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/book-club-class/count/{idvol}": {
			"get": {
				"operationId": "CountEvaluatedBookClubClassByIdVol",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"count": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"count"
									],
									"type": "object"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get total count of book club classes by a volunteer.",
				"tags": [
					"Book Club Class"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "idvol",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/book-club-class/available/{idvol}": {
			"get": {
				"operationId": "GetAvailableClasses",
				"responses": {
					"200": {
						"description": "Successfully fetched the essays",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AvailableClassRowEntity"
									},
									"type": "array"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get available essays for evaluation for the volunteer,\nthose essays includes the ones which does not have a reservation date\nor the reservations of the volunteer.\n\n(The volunteer must have bookPermission, which is checked using JWT)",
				"tags": [
					"Book Club Class"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "idvol",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/book-club-class/reservation": {
			"put": {
				"operationId": "ReserveClassForVolunteer",
				"responses": {
					"200": {
						"description": "Successfully reserved essay for volunteer",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AvailableClassRowEntity"
								}
							}
						}
					},
					"400": {
						"description": "Essay already reserved or evaluated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookClubClassError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "ESSAY_ALREADY_RESERVED_ERROR",
											"message": "Essay already reserved or already evaluated"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Essay not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookClubClassError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "ESSAY_NOT_FOUND",
											"message": "Essay with id {some class id} not found"
										}
									}
								}
							}
						}
					},
					"412": {
						"description": "Volunteer not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VolunteerError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "VOLUNTEER_NOT_FOUND",
											"message": "Volunteer with id {some volunteer id} not found"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Reserve essay for the volunteer. If the essay is already reserved or evaluated\nstatus 400 is returned.\n\n(The volunteer must have bookPermission, which is checked using JWT)",
				"tags": [
					"Book Club Class"
				],
				"security": [
					{
						"jwt": [
							"bookPermission"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReserveClassDataEntity"
							}
						}
					}
				}
			}
		},
		"/book-club-class/revert-reservation/{classId}": {
			"put": {
				"operationId": "RevertReserveClassForVolunteer",
				"responses": {
					"200": {
						"description": "Successfully reserved essay for volunteer",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AvailableClassRowEntity"
								}
							}
						}
					},
					"404": {
						"description": "Essay not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookClubClassError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "ESSAY_NOT_FOUND",
											"message": "Essay with id {some class id} not found"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Reserve essay for the volunteer.",
				"tags": [
					"Book Club Class"
				],
				"security": [
					{
						"jwt": [
							"bookPermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "classId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/book-club-class/from-id/{classId}": {
			"get": {
				"operationId": "GetClassesFromId",
				"responses": {
					"200": {
						"description": "Successfully fetched the classes",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AssociatedBCCEntity"
									},
									"type": "array"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all the book club classes starting from the classId in the path\n\n\n(The volunteer must have essayModulePermission, which is checked using JWT)",
				"tags": [
					"Book Club Class"
				],
				"security": [
					{
						"jwt": [
							"essayModulePermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "classId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/pep-class/from-id/{classId}": {
			"get": {
				"operationId": "GetClassesFromId",
				"responses": {
					"200": {
						"description": "Successfully fetched the classes",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PepClassEntity"
									},
									"type": "array"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all the pep classes starting from the classId in the path\n\n\n(The volunteer must have notebookModulePermission, which is checked using JWT)",
				"tags": [
					"Pep Class"
				],
				"security": [
					{
						"jwt": [
							"notebookModulePermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "classId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/pep-class/{classId}": {
			"put": {
				"operationId": "UpdateClass",
				"responses": {
					"200": {
						"description": "Successfully updated the class",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PepClassEntity"
								}
							}
						}
					},
					"400": {
						"description": "Could not update class",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PepClassError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "PEP_ClASS_NOT_UPDATED_ERROR",
											"message": "Class with ID {classId} not updated"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Could not find class",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PepClassError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "PEP_CLASS_NOT_FOUND",
											"message": "Class with id {classId} not found"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"details": {
											"$ref": "#/components/schemas/FieldErrors"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"details",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation Failed",
											"details": {
												"exampleField1": {
													"message": "invalid string value",
													"value": 12
												},
												"exampleField2": {
													"message": "'exampleField2' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Update class values available at UpdatePepClassEntity from class with classId\n\n\n(The volunteer must have notebookModulePermission, which is checked using JWT)",
				"tags": [
					"Pep Class"
				],
				"security": [
					{
						"jwt": [
							"notebookModulePermission"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "classId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePepClassEntity"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}